requirements.txt
absl-py
asgiref
astor
cached-property
Django
easydict
gast
google-pasta
grpcio
h5py
importlib-metadata
Keras-Applications
Keras-Preprocessing
Markdown
mysqlclient
numpy
opencv-python
opt-einsum
Pillow
protobuf
pytz
six
sqlparse
tensorboard
tensorflow
tensorflow-estimator
termcolor
typing-extensions
Werkzeug
wrapt
zipp
